# openapi.yaml
openapi: 3.1.0
info:
  title: VectorSpace Ad Server
  version: "0.1.0"
  description: |
    HTTP API for the VectorSpace ad exchange. This spec is generated from the
    wired routes in `handler/routes.go` and the actual request/response code in
    `handler/*.go`, `platform/`, `enclave/types.go`, and `trust/types.go` — not
    from the README (the README has drifted from the code; see the client notes
    below).

    Honest deployment notes:
    - **TEE / attestation is a placeholder in the default (mock) deployment.**
      `GET /tee/attestation` returns `attestation_cose_base64:
      "mock-attestation-not-for-production"` plus a real RSA public-key PEM. The
      default server runs `MockTEEProxy`, which decrypts the embedding **in the
      exchange's own process**. Do not rely on the TEE privacy property until
      real NSM attestation / vsock lands.
    - **`POST /ad-request` requires an encrypted envelope.** A plaintext
      `embedding` field is rejected with `400`. Encrypt against the public key
      from `GET /tee/attestation`.
    - **`POST /simulate` is the plaintext dev path** (admin-gated): plaintext
      `intent`, no logging, no billing. `POST /openrtb2/auction` is the other
      plaintext path (logged, billable), carrying the query the way OpenRTB
      does.

    Auth:
    - Admin endpoints require the `X-Admin-Password` header. If the server was
      started with an empty admin password (dev mode), the check is bypassed and
...

...
servers:
...
security: []
paths:
...
  /advertiser/register:
    post:
      tags: [advertiser]
      summary: Register an advertiser (intent position) or a keyword-import group
      description: |
        `intent` and `keywords` are mutually exclusive; exactly one is required.
        For intent positions, `sigma` must be > 0 (σ = 0 is the exact-match
        limit and is only valid for keyword imports, where it is the default).
        Returns a per-advertiser access token when a DB is configured.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/RegisterIntentResponse"
                  - $ref: "#/components/schemas/RegisterKeywordGroupResponse"
        "400": { $ref: "#/components/responses/PlainTextError" }
        "405": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
...
  /positions:
    get:
      tags: [advertiser]
      summary: List all registered positions
      description: |
        Returns the raw `Position` array. Note the `embedding` field is omitted
        from this serialization (JSON tag `-`); use `GET /embeddings` to fetch
        embeddings.
      security: []
      responses:
        "200":
          description: Positions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Position" }
        "405": { $ref: "#/components/responses/PlainTextError" }
...
  /embeddings:
    get:
      tags: [publisher]
      summary: All advertiser embeddings (ETag-cached, for SDK sync)
      security: []
      parameters:
        - name: If-None-Match
          in: header
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Embeddings
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                properties:
                  embeddings:
                    type: array
                    items: { $ref: "#/components/schemas/EmbeddingEntry" }
                required: [embeddings]
        "304": { description: Not Modified (ETag matched) }
        "405": { $ref: "#/components/responses/PlainTextError" }
  /embed:
    post:
      tags: [publisher]
      summary: Embed arbitrary text via the embedding sidecar
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text: { type: string }
              required: [text]
      responses:
        "200":
          description: Embedding vector
          content:
            application/json:
              schema:
                type: object
                properties:
                  embedding:
                    type: array
                    items: { type: number, format: double }
                required: [embedding]
        "400": { $ref: "#/components/responses/PlainTextError" }
        "405": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /simulate:
    post:
      tags: [auction]
      summary: Plaintext simulated auction (dev; no logging, no billing)
      description: |
        Admin-gated via `X-Admin-Password`. Includes ALL advertisers regardless
        of budget. If `tau` > 0, only positions within that squared-Euclidean
        distance of the query are included.
      security:
        - AdminPassword: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                intent: { type: string }
                tau: { type: number, format: double, default: 0 }
              required: [intent]
      responses:
        "200":
          description: Simulation result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimulationResult" }
        "400": { $ref: "#/components/responses/PlainTextError" }
        "401": { $ref: "#/components/responses/PlainTextError" }
        "405": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /chat:
    post:
      tags: [auction]
      summary: Proxy a chat completion to the Anthropic Messages API
      description: |
        Server-side proxy. Requires the server to have an Anthropic API key
        configured, otherwise returns `503`. Upstream API failures surface as
        `502`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                messages:
                  type: array
                  items:
                    type: object
                    properties:
                      role: { type: string, enum: [user, assistant] }
                      content: { type: string }
                    required: [role, content]
                system: { type: string }
              required: [messages]
      responses:
        "200":
          description: Assistant reply
          content:
            application/json:
              schema:
                type: object
                properties:
                  content: { type: string }
                required: [content]
        "400": { $ref: "#/components/responses/PlainTextError" }
        "405": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
        "502": { $ref: "#/components/responses/PlainTextError" }
        "503": { $ref: "#/components/responses/PlainTextError" }
  /openrtb2/auction:
    post:
      tags: [auction]
      summary: OpenRTB 2.5 wire-format auction (plaintext interop path)
      description: |
        Accepts a standard OpenRTB 2.5 BidRequest and returns a standard
        BidResponse. Two declared deviations from stock 2.5 semantics:
        settlement is per-click VCG (`bid.price` = the amount charged if the
        placement is clicked, `bid.ext.vectorspace.settlement = "cpc-vcg"`, and
        `imp.bidfloor` is per-click), and `adm` is a self-contained HTML snippet
        whose click-through routes through `GET /click` (where the charge
        fires). No-bid returns `204` with no body (ORTB convention). Query point
        precedence: `ext.vectorspace.embedding` → `ext.vectorspace.intent` →
        `site/app.content.keywords` → `user.keywords` (comma-separated, each
        embedded and scored separately). `test=1` requests are never logged and
        never billable.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ORTBBidRequest" }
      responses:
        "200":
          description: BidResponse
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ORTBBidResponse" }
        "204": { description: No bid (ORTB convention; empty body) }
        "400": { $ref: "#/components/responses/PlainTextError" }
        "405": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /ad-request:
    post:
      tags: [tee]
      summary: Private auction — encrypted embedding envelope (TEE path)
      description: |
        Receives an encrypted embedding, forwards it to the enclave, returns the
        auction outcome. A plaintext `embedding` field is rejected with `400`.
        `encrypted_embedding` requires `aes_key_encrypted`, `encrypted_payload`,
        and `nonce`; `hash_algorithm` is optional (defaults to SHA-256). The
        handler validates key and payload explicitly (`400`); a missing `nonce`
        currently surfaces at decryption time. In the default mock deployment the
        decrypt happens in-process (see TEE note in `info.description`).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuctionRequest" }
      responses:
        "200":
          description: Auction outcome
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdRequestResponse" }
        "400": { $ref: "#/components/responses/PlainTextError" }
        "405": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /tee/attestation:
    get:
      tags: [tee]
      summary: Enclave public key + attestation document
      description: |
        In the default mock deployment `attestation_cose_base64` is the literal
        placeholder `"mock-attestation-not-for-production"` and `public_key` is a
        real RSA PEM you can encrypt an `/ad-request` envelope against.
      security: []
      responses:
        "200":
          description: Attestation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AttestationResponse" }
        "405": { $ref: "#/components/responses/PlainTextError" }
        "503": { $ref: "#/components/responses/PlainTextError" }
  /event/impression:
    post:
      tags: [events]
      summary: Log an impression (frequency-capped when user_id is provided)
      description: Requires DB. Frequency-cap breach returns `429`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventRequest" }
      responses:
        "200": { $ref: "#/components/responses/StatusOk" }
        "400": { $ref: "#/components/responses/PlainTextError" }
        "405": { $ref: "#/components/responses/PlainTextError" }
        "429": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /event/click:
    post:
      tags: [events]
      summary: Log a click (charges the winner's budget on first click)
      description: |
        Requires DB. `auction_id` must be a real auction (> 0) whose winner
        matches `advertiser_id`, else `400`. If the first-click budget charge
        fails (budget missing or exhausted) the response is `402`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventRequest" }
      responses:
        "200": { $ref: "#/components/responses/StatusOk" }
        "400": { $ref: "#/components/responses/PlainTextError" }
        "402": { $ref: "#/components/responses/PlainTextError" }
        "405": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /event/viewable:
    post:
      tags: [events]
      summary: Log a viewability event
      description: Requires DB.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventRequest" }
      responses:
        "200": { $ref: "#/components/responses/StatusOk" }
        "400": { $ref: "#/components/responses/PlainTextError" }
        "405": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /click:
    get:
      tags: [events]
      summary: Settlement redirect (charge on first click, then 302 to winner URL)
      description: |
        Requires DB. Charges the winner's budget on the first click for the
        auction, logs it, then redirects (`302`) to the winner's URL. If the
        winner has no URL, returns `200` with `{"status":"ok"}` instead.
      security: []
      parameters:
        - name: auction_id
          in: query
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
      responses:
        "200": { $ref: "#/components/responses/StatusOk" }
        "302": { description: Redirect to the winning advertiser's URL }
        "400": { $ref: "#/components/responses/PlainTextError" }
        "402": { $ref: "#/components/responses/PlainTextError" }
        "404": { $ref: "#/components/responses/PlainTextError" }
        "405": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /portal/me:
    get:
      tags: [portal-advertiser]
      summary: Advertiser profile + budget
      security:
        - PortalToken: []
      parameters:
        - $ref: "#/components/parameters/Token"
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PortalMe" }
        "401": { $ref: "#/components/responses/PlainTextError" }
        "404": { $ref: "#/components/responses/PlainTextError" }
    put:
      tags: [portal-advertiser]
      summary: Update advertiser profile
      security:
        - PortalToken: []
      parameters:
        - $ref: "#/components/parameters/Token"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateRequest" }
      responses:
        "200":
          description: Updated position
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Position" }
        "400": { $ref: "#/components/responses/PlainTextError" }
        "401": { $ref: "#/components/responses/PlainTextError" }
        "404": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /portal/me/auctions:
    get:
      tags: [portal-advertiser]
      summary: Auction history for the token's advertiser (JSON or CSV)
      security:
        - PortalToken: []
      parameters:
        - $ref: "#/components/parameters/Token"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: format
          in: query
          required: false
          description: Set to `csv` for a CSV attachment (ignores limit/offset).
          schema: { type: string, enum: [csv] }
      responses:
        "200":
          description: Auctions (JSON) or CSV attachment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuctionsPage" }
            text/csv:
              schema: { type: string }
        "401": { $ref: "#/components/responses/PlainTextError" }
        "405": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /portal/me/events:
    get:
      tags: [portal-advertiser]
      summary: Event stats for the token's advertiser
      security:
        - PortalToken: []
      parameters:
        - $ref: "#/components/parameters/Token"
      responses:
        "200":
          description: Event stats
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventStats" }
        "401": { $ref: "#/components/responses/PlainTextError" }
        "405": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /portal/me/creatives:
    get:
      tags: [portal-advertiser]
      summary: List creatives for the token's advertiser
      security:
        - PortalToken: []
      parameters:
        - $ref: "#/components/parameters/Token"
      responses:
        "200":
          description: Creatives
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Creative" }
        "401": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
    post:
      tags: [portal-advertiser]
      summary: Create a creative
      security:
        - PortalToken: []
      parameters:
        - $ref: "#/components/parameters/Token"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                subtitle: { type: string }
              required: [title]
      responses:
        "201":
          description: Created creative
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Creative" }
        "400": { $ref: "#/components/responses/PlainTextError" }
        "401": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
...
  /portal/publisher/me:
    get:
      tags: [portal-publisher]
      summary: Publisher profile
      security:
        - PortalToken: []
      parameters:
        - $ref: "#/components/parameters/Token"
      responses:
        "200":
          description: Publisher
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Publisher" }
        "401": { $ref: "#/components/responses/PlainTextError" }
        "404": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /portal/publisher/stats:
    get:
      tags: [portal-publisher]
      summary: Aggregate stats for the token's publisher
      security:
        - PortalToken: []
      parameters:
        - $ref: "#/components/parameters/Token"
      responses:
        "200":
          description: Publisher stats
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublisherStats" }
        "401": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /portal/publisher/revenue:
    get:
      tags: [portal-publisher]
      summary: Revenue over time for the token's publisher
      security:
        - PortalToken: []
      parameters:
        - $ref: "#/components/parameters/Token"
        - $ref: "#/components/parameters/GroupBy"
      responses:
        "200":
          description: Revenue periods
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RevenuePage" }
        "401": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /portal/publisher/events:
    get:
      tags: [portal-publisher]
      summary: Event stats for the token's publisher
      security:
        - PortalToken: []
      parameters:
        - $ref: "#/components/parameters/Token"
      responses:
        "200":
          description: Event stats
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventStats" }
        "401": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /portal/publisher/auctions:
    get:
      tags: [portal-publisher]
      summary: Auction history for the token's publisher
      security:
        - PortalToken: []
      parameters:
        - $ref: "#/components/parameters/Token"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Auctions page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuctionsPage" }
        "401": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /portal/publisher/top-advertisers:
    get:
      tags: [portal-publisher]
      summary: Top advertisers by spend on the token's publisher
      security:
        - PortalToken: []
      parameters:
        - $ref: "#/components/parameters/Token"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Advertiser spend list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdvertiserSpend" }
        "401": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /publisher/register:
    post:
      tags: [publisher]
      summary: Register a publisher (admin-protected)
      description: Requires DB. Returns a publisher token.
      security:
        - AdminPassword: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                domain: { type: string }
              required: [name]
      responses:
        "201":
          description: Created publisher
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, example: pub-1 }
                  name: { type: string }
                  domain: { type: string }
                  token: { type: string }
                required: [id, name, domain, token]
        "400": { $ref: "#/components/responses/PlainTextError" }
        "401": { $ref: "#/components/responses/PlainTextError" }
        "405": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /publisher/login:
    post:
      tags: [publisher]
      summary: Authenticate a publisher by email/password
      description: Requires DB. Returns the publisher's token on success.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string }
              required: [email, password]
      responses:
        "200":
          description: Token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  publisher_id: { type: string }
                required: [token, publisher_id]
        "400": { $ref: "#/components/responses/PlainTextError" }
        "401": { $ref: "#/components/responses/PlainTextError" }
        "405": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /admin/auctions:
    get:
      tags: [admin]
      summary: All auctions (audit trail; JSON or CSV)
      security:
        - AdminPassword: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: winner
          in: query
          required: false
          schema: { type: string }
        - name: intent
          in: query
          required: false
          schema: { type: string }
        - name: format
          in: query
          required: false
          schema: { type: string, enum: [csv] }
      responses:
        "200":
          description: Auctions page (JSON) or CSV attachment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuctionsPage" }
            text/csv:
              schema: { type: string }
        "401": { $ref: "#/components/responses/PlainTextError" }
        "405": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
  /admin/revenue:
    get:
      tags: [admin]
      summary: Revenue analytics (grouped by day/week/month)
      security:
        - AdminPassword: []
      parameters:
        - $ref: "#/components/parameters/GroupBy"
      responses:
        "200":
          description: Revenue periods
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RevenuePage" }
        "401": { $ref: "#/components/responses/PlainTextError" }
        "405": { $ref: "#/components/responses/PlainTextError" }
        "500": { $ref: "#/components/responses/PlainTextError" }
...
components:
  securitySchemes:
    AdminPassword:
      type: apiKey
      in: header
      name: X-Admin-Password
      description: |
        Admin password header. If the server was started with an empty admin
        password (dev mode), the check is skipped and any request passes.
    PortalToken:
      type: apiKey
      in: query
      name: token
      description: |
        Advertiser or publisher access token, passed as the `token` query
        parameter. Advertiser tokens come from `/advertiser/register`; publisher
        tokens from `/publisher/register`, `/admin/publishers`, or
        `/publisher/login`.
...